+++ src/Adapters/BlinkerSerialSIM7020.h
@@ -61,7 +61,7 @@ class BlinkerSerialSIM7020 : public BlinkerStream
         void duerType(const String & type);
         void begin(const char* _deviceType, String _imei);
         void initStream(Stream& s, bool state, blinker_callback_t func);
-        int autoPrint(uint32_t id);
+        int autoPrint(unsigned long id);
         char * deviceName();
         char * authKey() { return AUTHKEY_NBIoT; }
         int init() { return isMQTTinit; }
@@ -817,7 +817,7 @@ void BlinkerSerialSIM7020::initStream(Stream& s, bool state, blinker_callback_t 
     // strcpy(_imei, imei.c_str());
 }
 
-int BlinkerSerialSIM7020::autoPrint(uint32_t id)
+int BlinkerSerialSIM7020::autoPrint(unsigned long id)
 {
     String payload = BLINKER_F("{\"data\":{\"set\":{");
     payload += BLINKER_F("\"trigged\":true,\"autoData\":{");
